refactor(BE; Dimensions): WIP delete selectable Option

+++ diversity-lunch-be/src/main/java/de/brockhausag/diversitylunchspringboot/dimensions/services/model/BasicDimensionService.java
@@ -57,8 +57,12 @@ public class BasicDimensionService implements DimensionService<
     }
 
     @Override
-    public void deleteSelectableOptionById(Long selectableOptionId) {
+    public boolean deleteSelectableOptionById(Long selectableOptionId) {
+        if (!selectableRepository.existsById(selectableOptionId)) {
+            return false;
+        }
         selectableRepository.deleteById(selectableOptionId);
+        return !selectableRepository.existsById(selectableOptionId);
     }
 
     @Override
