+++ diversity-lunch-be/src/main/java/de/brockhausag/diversitylunchspringboot/dimensions/services/model/MultiselectDimensionService.java
@@ -39,8 +39,12 @@ public class MultiselectDimensionService implements DimensionService<Multiselect
     }
 
     @Override
-    public void deleteSelectableOptionById(Long selectableOptionId) {
+    public boolean deleteSelectableOptionById(Long selectableOptionId) {
+        if (!selectableRepository.existsById(selectableOptionId)) {
+            return false;
+        }
         selectableRepository.deleteById(selectableOptionId);
+        return !selectableRepository.existsById(selectableOptionId);
     }
 
     @Override
